Force casks to reinstall if Homebrew gets removed
When running Ruby on Mac in reset mode, it gets rid of the Caskroom, but it still leaves the apps in /Applications. But then if you want to restore everything, Homebrew will fail to install the casks because it sees the app already exists. By setting force to true, it reinstalls the apps using Cask.